회원 수정 삭제 등 에서 권한 검증

+++ src/main/java/com/example/prj1be/controller/BoardController.java
@@ -54,7 +54,15 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+        @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
@@ -63,8 +71,16 @@ public class BoardController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Board board) {
+    public ResponseEntity edit(@RequestBody Board board,
+        @SessionAttribute(value = "login", required = false) Member login) {
 //        System.out.println("board = " + board);
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(board.getId(), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         if (service.validate(board)) {
 
             if (service.update(board)) {
